Moved around user confirmation personalisation  for readability
